backend/requirements.txt
# Reflektionsarkiv backend – fasta versioner för determinism
fastapi==0.115.6
uvicorn[standard]==0.32.1
mysql-connector-python==9.0.0
pydantic==2.10.3
pydantic-settings==2.6.1
openai==1.55.3
httpx==0.27.2
pytest==8.3.4
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.10.1
